Fix compilation of nativebridge on Android S
Use LOCAL_HEADERS_LIBRARIES instead of hard-coded include paths.

Move ro.dalvik.vm.native.bridge to PRODUCT_PRODUCT_PROPERTIES. As we're
not using a vendor image, PRODUCT_PROPERTY_OVERRIDES content ends up in
/system/build.prop. However, Android S now forbids duplicate sysprop
entries on the same partition (unless they're marked optional, but
that's not the case here). Simple solution is to move this sysprop to
/product/etc/build.prop.